fix bugs for create and delete
