+++ front/src/Components/App.js
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from "react";
 import { BrowserRouter, NavLink, Routes, Route } from "react-router-dom";
 import Signin from "./signin";
 import Signup from "./signup";
-import Transaction from "./transaction";
 import Stat from "./stat";
 import Dashboard from "./dashboard";
 import logo from "../images/logo.png";
@@ -10,14 +9,13 @@ import "../stylesheets/style.css";
 
 export default function App() {
   const [user, setUser] = useState();
-  const [flag, refreshPage] = useState(true);
 
   useEffect(() => {
     getUsername().then((username) => {
       setUser(username);
       console.log(username);
     });
-  }, [flag]);
+  }, []);
 
   return (
     <BrowserRouter>
@@ -56,13 +54,12 @@ export default function App() {
               <Routes>
                 <Route path="/signin" element={<Signin />} />
                 <Route path="/signup" element={<Signup />} />
-                <Route path="/transaction" element={<Transaction />} />
                 <Route path="/stat" element={<Stat />} />
                 <Route path="/dashboard" element={<Dashboard />} />
               </Routes>
             </div>
           ) : (
-            <Dashboard username={user} refreshPage={refreshPage} />
+            <Dashboard username={user} />
           )}
         </div>
       </div>
